renamed class to make more sense
