+++ examples/post-type-tax-map.json
@@ -2,13 +2,5 @@
   {
     "from": "category",
     "to": "event-category"
-  },
-  {
-    "from": "invalid-category",
-    "to": "event-category"
-  },
-  {
-    "from": "category",
-    "to": "invalid-category"
   }
 ]
